fix: カスタムフックの見直し
