+++ src/hooks/useUser.ts
@@ -1,36 +1,45 @@
-import { useState, useCallback, useMemo } from "react";
+import { useMemo } from "react";
+import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { User } from "@supabase/supabase-js";
 import { createClient } from "@/lib/supabase/supabase";
 
-const useUser = () => {
-  const [user, setUser] = useState<User | null>(null);
-  const [loading, setLoading] = useState(true);
+export const USER_QUERY_KEY = ["user"] as const;
 
+const useUser = (): {
+  user: User | null | undefined;
+  loading: boolean;
+  signOut: () => Promise<void>;
+} => {
   const supabase = useMemo(() => createClient(), []);
+  const queryClient = useQueryClient();
 
-  const fetchUser = useCallback(async () => {
-    setLoading(true);
-    try {
+  const { data: user, isLoading: loading } = useQuery({
+    queryKey: USER_QUERY_KEY,
+    queryFn: async () => {
       const {
-        data: { user: supabaseUser },
+        data: { user },
+        error,
       } = await supabase.auth.getUser();
+      if (error) {
+        throw error;
+      }
+      return user ?? null;
+    },
+  });
 
-      setUser(supabaseUser);
-      return supabaseUser;
-    } finally {
-      setLoading(false);
-    }
-  }, [supabase]);
+  const { mutateAsync: signOut } = useMutation({
+    mutationFn: async () => {
+      const { error } = await supabase.auth.signOut();
+      if (error) {
+        throw error;
+      }
+    },
+    onSuccess: () => {
+      queryClient.setQueryData(USER_QUERY_KEY, null);
+    },
+  });
 
-  const signOut = useCallback(async () => {
-    const { error } = await supabase.auth.signOut();
-    if (error) {
-      throw error;
-    }
-    setUser(null);
-  }, [supabase]);
-
-  return { user, loading, fetchUser, signOut };
+  return { user, loading, signOut };
 };
 
 export default useUser;
